finished moving things around and resolving imports
